Swithcing translation to be via gemini
